split to dedicated wasm files
